EN-1970 Fix tests for ApiClient.
